MainActivity, RetroFit, ExerciseImage, etc.
The MainActivity file, as of now, is displaying Strings of the image URL (Needs fixing). The RetroFit class is where we define our base URL. The ExerciseImage file holds the variables that are a part of our API along with the getters and setters for easier accessibility. Lastly, the ExerciseImageAPI is the interface that holds our relative URL